backend/requirements.txt
fastapi
uvicorn
spotipy
pandas
python-dotenv
httpx
requests
librosa
av
numpy
scipy
soundfile
pydub
ffmpeg-python
decorator
pyarrow
scikit-learn
tqdm
datasets
